requirements.txt
Django==4.2.10
djangorestframework==3.14.0
celery==5.3.6
redis==5.0.1
stripe==7.10.0
twilio==8.12.0
python-dotenv==1.0.1